Refactor PaginatedList to Domain/Contracts
