+++ API/BitsionTest.API/Domain/Contracts/PaginatedList.cs
@@ -1,0 +1,18 @@
+﻿namespace BitsionTest.API.Domain.Contracts
+{
+    public class PaginatedList<T>
+    {
+        public List<T> Items { get; set; }
+        public int TotalCount { get; set; } // total de clientes
+        public int PageNumber { get; set; }
+        public int PageSize { get; set; }
+
+        public PaginatedList(List<T> items, int totalCount, int pageNumber, int pageSize)
+        {
+            Items = items;
+            TotalCount = totalCount;
+            PageNumber = pageNumber;
+            PageSize = pageSize;
+        }
+    }
+}
