+++ API/BitsionTest.API/Repositories/Implementation/ClientRepository.cs
@@ -1,4 +1,5 @@
-﻿using BitsionTest.API.Domain.Entities;
+﻿using BitsionTest.API.Domain.Contracts;
+using BitsionTest.API.Domain.Entities;
 using BitsionTest.API.Infrastructure.Context;
 using BitsionTest.API.Repositories.Interface;
 using Microsoft.EntityFrameworkCore;
@@ -109,19 +110,4 @@ public class ClientRepository : IClientRepository
     }
 
 
-    public class PaginatedList<T>
-    {
-        public List<T> Items { get; set; }
-        public int TotalCount { get; set; } // total de clientes
-        public int PageNumber { get; set; }
-        public int PageSize { get; set; }
-
-        public PaginatedList(List<T> items, int totalCount, int pageNumber, int pageSize)
-        {
-            Items = items;
-            TotalCount = totalCount;
-            PageNumber = pageNumber;
-            PageSize = pageSize;
-        }
-    }
 }
